refactor: reorganize deployment scripts and update import paths
- Updated import paths across deployment scripts to reflect a new common structure, enhancing modularity and clarity.
- Consolidated utility functions related to fleet and ark deployments into dedicated modules for better reusability.
- Introduced new scripts for managing bridge configurations and governance proposals, improving overall maintainability.
- Removed deprecated files and streamlined the deployment process, focusing on essential functionalities.
