+++ packages/deployment/scripts/arks/deploy-origineth-ark.ts
@@ -9,9 +9,9 @@ import {
 } from '../../ignition/modules/arks/origineth-ark'
 import { ArkParams } from '../../types/ark-params'
 import { BaseConfig, Token } from '../../types/config-types'
-import { BaseArkParams } from '../arks/core/deployment'
+import { BaseArkParams } from './deployment'
 import { HUNDRED_PERCENT, MAX_UINT256_STRING } from '../common/constants'
-import { getFleetConfig } from '../fleets/deployment/config'
+import { getFleetConfig } from '../fleets/config'
 import { handleDeploymentId } from '../common/files'
 import { getChainId } from '../common/chain'
 import { continueDeploymentCheck } from '../common/prompts'
