+++ packages/deployment/scripts/arks/deploy-silo-ark-v2.ts
@@ -4,9 +4,9 @@ import prompts from 'prompts'
 import { Address } from 'viem'
 import { createSiloArkV2Module, SiloArkV2Contracts } from '../../ignition/modules/arks/silo-ark-v2'
 import { BaseConfig, Token } from '../../types/config-types'
-import { BaseArkParams } from '../arks/core/deployment'
+import { BaseArkParams } from './deployment'
 import { HUNDRED_PERCENT, MAX_UINT256_STRING } from '../common/constants'
-import { getFleetConfig } from '../fleets/deployment/config'
+import { getFleetConfig } from '../fleets/config'
 import { handleDeploymentId } from '../common/files'
 import { getChainId } from '../common/chain'
 import { continueDeploymentCheck } from '../common/prompts'
