+++ packages/deployment/scripts/arks/deploy-silo-ark.ts
@@ -4,10 +4,10 @@ import prompts from 'prompts'
 import { Address } from 'viem'
 import { createSiloArkModule, SiloArkContracts } from '../../ignition/modules/arks/silo-ark'
 import { BaseConfig, Token } from '../../types/config-types'
-import { BaseArkParams } from '../arks/core/deployment'
+import { BaseArkParams } from './deployment'
 import { getChainId } from '../common/chain'
 import { HUNDRED_PERCENT, MAX_UINT256_STRING } from '../common/constants'
-import { getFleetConfig } from '../fleets/deployment/config'
+import { getFleetConfig } from '../fleets/config'
 import { handleDeploymentId } from '../common/files'
 import { continueDeploymentCheck } from '../common/prompts'
