+++ packages/deployment/scripts/arks/deploy-sky-rewards-ark.ts
@@ -8,7 +8,7 @@ import {
 } from '../../ignition/modules/arks/sky-rewards-ark'
 import { BaseConfig, Token } from '../../types/config-types'
 import { ADDRESS_ZERO, HUNDRED_PERCENT, MAX_UINT256_STRING } from '../common/constants'
-import { getFleetConfig } from '../fleets/deployment/config'
+import { getFleetConfig } from '../fleets/config'
 import { handleDeploymentId } from '../common/files'
 import { getChainId } from '../common/chain'
 import { continueDeploymentCheck } from '../common/prompts'
