+++ packages/deployment/scripts/arks/deploy-spark-ark.ts
@@ -4,9 +4,9 @@ import prompts from 'prompts'
 import { Address } from 'viem'
 import { SparkArkContracts, createSparkArkModule } from '../../ignition/modules/arks/spark-ark'
 import { BaseConfig, Token } from '../../types/config-types'
-import { BaseArkParams } from '../arks/core/deployment'
+import { BaseArkParams } from './deployment'
 import { HUNDRED_PERCENT, MAX_UINT256_STRING } from '../common/constants'
-import { getFleetConfig } from '../fleets/deployment/config'
+import { getFleetConfig } from '../fleets/config'
 import { handleDeploymentId } from '../common/files'
 import { getChainId } from '../common/chain'
 import { continueDeploymentCheck } from '../common/prompts'
