+++ packages/deployment/scripts/governance/bridge/configure-new-chain-lz.ts
@@ -3,14 +3,14 @@ import kleur from 'kleur'
 import prompts from 'prompts'
 import { Address, createWalletClient, encodeAbiParameters, http } from 'viem'
 import { privateKeyToAccount } from 'viem/accounts'
-import { CHAIN_CONFIG_MAP, RPC_URL_MAP } from '../../core/chain'
-import { getChainPublicClient } from '../../core/clients'
-import { getConfigByNetwork } from '../../core/config'
-import { promptForFleetDeploymentOutput } from '../../fleets/core/deployment'
-import { promptForConfigType } from '../../utils/prompts'
-import { warnIfTenderlyVirtualTestnet } from '../../utils/tenderly'
-import { createUnifiedLzConfigProposal } from './common/bridge-governance-helper'
-import { checkLzAuthorization } from './common/lz-authorization-helper'
+import { CHAIN_CONFIG_MAP, RPC_URL_MAP } from '../../common/chain'
+import { getChainPublicClient } from '../../common/clients'
+import { getConfigByNetwork } from '../../common/config'
+import { promptForConfigType } from '../../common/prompts'
+import { warnIfTenderlyVirtualTestnet } from '../../common/tenderly'
+import { promptForFleetDeploymentOutput } from '../../fleets/deployment'
+import { createUnifiedLzConfigProposal } from './bridge-governance'
+import { checkLzAuthorization } from './lz-authorization'
 import { LZ_ENDPOINT_ABI } from './lz-endpoint-abi'
 
 // Interface for LayerZero configuration
