+++ packages/deployment/scripts/governance/execute-proposal.ts
@@ -2,12 +2,12 @@ import hre from 'hardhat'
 import kleur from 'kleur'
 import prompts from 'prompts'
 import { Address, parseAbi } from 'viem'
-import { HUB_CHAIN_NAME } from '../common/constants'
 import { getConfigByNetwork } from '../common/config'
+import { HUB_CHAIN_NAME } from '../common/constants'
 import { hashDescription } from '../common/hash-description'
 import { promptForConfigType } from '../common/prompts'
-import { ProposalData, loadProposal } from '../governance/core/proposal'
 import { sleep } from '../common/utils'
+import { ProposalData, loadProposal } from '../governance/proposal'
 
 enum ProposalState {
   Pending,
